update perf suite to match rustc_offload_frontend style
